feat(reducer): extract source-agnostic diagnostic core
