+++ crates/bazel-mcp-reducer-cases/Cargo.toml
@@ -15,6 +15,7 @@ bazel-mcp-bep.workspace = true
 bazel-mcp-reducer.workspace = true
 bazel-mcp-store.workspace = true
 bazel-mcp-types.workspace = true
+diagnostic-reducer.workspace = true
 clap.workspace = true
 regex.workspace = true
 schemars.workspace = true
