+++ crates/diagnostic-reducer/src/diagnostics/common.rs
